+++ src/paddlegeometry.cpp
@@ -52,7 +52,7 @@ void PaddleGeometry::Init(Renderer* renderer, float arcAngle, int arcDivisions, 
     vec2 endForwardBasis = vec2(-endLeftBasis.y, endLeftBasis.x);
     
     const float startCapAngle = 0.0f;
-    const float capAngleSteps = M_PI / capDivisions;
+    const float capAngleSteps = (float)M_PI / capDivisions;
 
     for (int i = 0; i < capDivisions - 1; ++i)
     {
@@ -71,7 +71,7 @@ void PaddleGeometry::Init(Renderer* renderer, float arcAngle, int arcDivisions, 
     const int numIndices = numArcIndices + numCapIndices;
     std::vector<short> indices(numIndices);
 
-    for (int i = 0; i < arcDivisions - 1; ++i)
+    for (short i = 0; i < (short)arcDivisions - 1; ++i)
     {
         short a0 = i;
         short a1 = i + (short)arcDivisions;
@@ -96,35 +96,35 @@ void PaddleGeometry::Init(Renderer* renderer, float arcAngle, int arcDivisions, 
         indices[++o] = a2;
     }
 
-    short startCapMid = arcDivisions;
+    short startCapMid = (short)arcDivisions;
     short startCapPrev = 0;
-    short startCapIndex = numArcIndices;
+    short startCapIndex = (short)numArcIndices;
 
-    short endCapMid = arcDivisions * 2 - 1;
-    short endCapPrev = arcDivisions - 1;
-    short endCapIndex = numArcIndices + capDivisions * 3;
+    short endCapMid = (short)arcDivisions * 2 - 1;
+    short endCapPrev = (short)arcDivisions - 1;
+    short endCapIndex = (short)(numArcIndices + capDivisions * 3);
 
     for (int i = 0; i < capDivisions - 1; ++i)
     {
-        short startCap = numArcVerts + i;
+        short startCap = short(numArcVerts + i);
         indices[startCapIndex++] = startCap;
         indices[startCapIndex++] = startCapPrev;
         indices[startCapIndex++] = startCapMid;
         startCapPrev = startCap;
 
-        short endCap = numArcVerts + (capDivisions - 1) + i;
+        short endCap = (short)(numArcVerts + (capDivisions - 1) + i);
         indices[endCapIndex++] = endCapPrev;
         indices[endCapIndex++] = endCap;
         indices[endCapIndex++] = endCapMid;
         endCapPrev = endCap;
     }
 
-    indices[startCapIndex++] = arcDivisions * 2;
+    indices[startCapIndex++] = (short)arcDivisions * 2;
     indices[startCapIndex++] = startCapPrev;
     indices[startCapIndex++] = startCapMid;
 
     indices[endCapIndex++] = endCapPrev;
-    indices[endCapIndex++] = arcDivisions * 3 - 1;
+    indices[endCapIndex++] = (short)arcDivisions * 3 - 1;
     indices[endCapIndex++] = endCapMid;
 
     m_vertbuffer = renderer->CreateBuffer(verts.size() * sizeof(Paddle_Vert), verts.data());
